Refactor file structure (#257)
* rename OMParser/__init__.py => OMParser.py

reason: cleanup; no need to use a sub-module / sub-directory

* prepare new file ModelicaSystem.py

* dummy entries for the classes to move from __init__.py

* prepare new file OMsession.py

* dummy entries for the classes to move from __init__.py

* prepare new file OMsession.py (2)

* fix imports based on merged changes

* move class content to the new files

!! only copy & paste of the class definitions - *NO* further change !!

* OMCSessionBase => OMCSession.py
* OMCSessionZMQ => OMCSession.py
* ModelicaSystem => ModelicaSystem.py
* ModelicaSystemError => ModelicaSystem.py

* cleanup of __init__.py after moving all classes into separate files

* fix pyproject.toml - remove required subdirectory OMPython/OMParser

* move license to the top of the file
